ssmt.h: extract NAN/v8 data converter methods
* Focus ssmt.h on only steam-specific items, not general functions.
* Enables reuse - these methods are duplicated in other files.
